refactor: divide card move functions from main and make card dir for card actions
